Change module names and structure for mpec.
